requirements.txt

arm-pyart==1.10.2
boto3==1.9.236
matplotlib==3.1.1
netCDF4==1.5.2
numpy==1.17.2
scipy==1.3.1
xarray==0.13.0

#Cython==0.29.13
#Shapely==1.6.4.post2
#Cartopy==0.17.0

